+++ packages/algob/src/internal/deployer.ts
@@ -1,11 +1,16 @@
-import { overrideASADef, parseASADef, types as rtypes, validateOptInAccNames } from "@algo-builder/runtime";
+import {
+  overrideASADef,
+  parseASADef,
+  types as rtypes,
+  validateOptInAccNames
+} from "@algo-builder/runtime";
 import { BuilderError, ERRORS, types as wtypes } from "@algo-builder/web";
 import type { EncodedMultisig, LogicSigAccount, modelsv2 } from "algosdk";
 import * as algosdk from "algosdk";
-import { cloneDeep } from "lodash";
 
 import { txWriter } from "../internal/tx-log-writer";
 import { AlgoOperator } from "../lib/algo-operator";
+import { CompileOp } from "../lib/compile";
 import { getDummyLsig, getLsig, getLsigFromCache } from "../lib/lsig";
 import { blsigExt, loadBinaryLsig, readMsigFromFile } from "../lib/msig";
 import { CheckpointFunctionsImpl, persistCheckpoint } from "../lib/script-checkpoints";
@@ -43,7 +48,10 @@ class DeployerBasicMode {
     this.accounts = deployerCfg.runtimeEnv.network.config.accounts;
     this.accountsByName = deployerCfg.accounts;
     this.txWriter = deployerCfg.txWriter;
-    this.checkpoint = new CheckpointFunctionsImpl(deployerCfg.cpData, deployerCfg.runtimeEnv.network.name);
+    this.checkpoint = new CheckpointFunctionsImpl(
+      deployerCfg.cpData,
+      deployerCfg.runtimeEnv.network.name
+    );
     this.indexerClient = deployerCfg.indexerClient;
   }
 
@@ -58,10 +66,9 @@ class DeployerBasicMode {
   getASAInfo (name: string): rtypes.ASAInfo {
     const found = this.asa.get(name);
     if (!found) {
-      throw new BuilderError(
-        ERRORS.BUILTIN_TASKS.DEPLOYER_ASA_NOT_DEFINED, {
-          assetName: name
-        });
+      throw new BuilderError(ERRORS.BUILTIN_TASKS.DEPLOYER_ASA_NOT_DEFINED, {
+        assetName: name
+      });
     }
     return found;
   }
@@ -69,10 +76,9 @@ class DeployerBasicMode {
   private _getAccount (name: string): rtypes.Account {
     const found = this.accountsByName.get(name);
     if (!found) {
-      throw new BuilderError(
-        ERRORS.BUILTIN_TASKS.ACCOUNT_NOT_FOUND, {
-          assetName: name
-        });
+      throw new BuilderError(ERRORS.BUILTIN_TASKS.ACCOUNT_NOT_FOUND, {
+        assetName: name
+      });
     }
     return found;
   }
@@ -155,21 +161,25 @@ class DeployerBasicMode {
    * Loads a single signed delegated logic signature account from checkpoint
    */
   getDelegatedLsig (lsigName: string): LogicSigAccount | undefined {
-    const resultMap = this.cpData.precedingCP[this.networkName]?.dLsig ?? new Map(); ;
+    const resultMap = this.cpData.precedingCP[this.networkName]?.dLsig ?? new Map();
     const result = resultMap.get(lsigName)?.lsig;
-    if (result === undefined) { return undefined; }
+    if (result === undefined) {
+      return undefined;
+    }
 
     const lsigAccount = Object.assign(getDummyLsig(), result);
     lsigAccount.lsig = Object.assign(getDummyLsig().lsig, result.lsig);
 
-    if (lsigAccount.lsig.sig) { lsigAccount.lsig.sig = Uint8Array.from(lsigAccount.lsig.sig); };
+    if (lsigAccount.lsig.sig) {
+      lsigAccount.lsig.sig = Uint8Array.from(lsigAccount.lsig.sig);
+    }
     return lsigAccount;
   }
 
   /**
    * Loads a logic signature account from checkpoint
    * @param lsigName logic signature name
-  */
+   */
   getContractLsig (lsigName: string): LogicSigAccount | undefined {
     return this.getDelegatedLsig(lsigName);
   }
@@ -195,16 +205,51 @@ class DeployerBasicMode {
   }
 
   /**
+   * Alias to `this.compileASC` with last two parameters being swapped.
+   * Deprecated: this function will be removed in the next release.
+   */
+  ensureCompiled (
+    name: string,
+    force?: boolean,
+    scTmplParams?: SCParams
+  ): Promise<ASCCache> {
+    return this.compileASC(name, scTmplParams, force);
+  }
+
+  /**
+   * Returns ASCCache (with compiled code)
+   * @param name: Smart Contract filename (must be present in assets folder)
+   * @param scTmplParams: scTmplParams: Smart contract template parameters
+   *     (used only when compiling PyTEAL to TEAL)
+   * @param force: if force is true file will be compiled for sure, even if it's checkpoint exist
+   */
+  compileASC (name: string, scTmplParams?: SCParams, force?: boolean): Promise<ASCCache> {
+    return this.algoOp.ensureCompiled(name, force, scTmplParams);
+  }
+
+  /**
+   * Returns cached program (from artifacts/cache) `ASCCache` object by filename.
+   * TODO: beta support - this will change
+   * @param name ASC name used during deployment
+   */
+  getDeployedASC (name: string): Promise<ASCCache | undefined> {
+    const op = new CompileOp(this.algoOp.algodClient);
+    return op.readArtifact(name);
+  }
+
+  /**
    * Loads multisigned logic signature account from .lsig or .blsig file
    * @param name filename
    * @returns multi signed logic signature from assets/<file_name>.(b)lsig
    */
   async loadMultiSig (name: string): Promise<LogicSig> {
-    if (name.endsWith(blsigExt)) { return await loadBinaryLsig(name); }
+    if (name.endsWith(blsigExt)) {
+      return await loadBinaryLsig(name);
+    }
 
     const lsig = (await getLsig(name, this.algoOp.algodClient)).lsig; // get lsig from .teal (getting logic part from lsig)
     const msig = await readMsigFromFile(name); // Get decoded Msig object from .msig
-    Object.assign(lsig.msig = {} as EncodedMultisig, msig);
+    Object.assign((lsig.msig = {} as EncodedMultisig), msig);
     return lsig;
   }
 
@@ -212,10 +257,8 @@ class DeployerBasicMode {
    * Send signed transaction to network and wait for confirmation
    * @param rawTxns Signed Transaction(s)
    */
-  async sendAndWait (
-    rawTxns: Uint8Array | Uint8Array[]
-  ): Promise<ConfirmedTxInfo> {
-    return await this.algoOp.sendAndWait(rawTxns);
+  sendAndWait (rawTxns: Uint8Array | Uint8Array[]): Promise<ConfirmedTxInfo> {
+    return this.algoOp.sendAndWait(rawTxns);
   }
 
   /**
@@ -225,7 +268,11 @@ class DeployerBasicMode {
    * @param accountName
    * @param flags Transaction flags
    */
-  async optInAccountToASA (asa: string, accountName: string, flags: wtypes.TxParams): Promise<void> {
+  optInAccountToASA (
+    asa: string,
+    accountName: string,
+    flags: wtypes.TxParams
+  ): Promise<void> {
     this.assertCPNotDeleted({
       type: wtypes.TransactionType.OptInASA,
       sign: wtypes.SignType.SecretKey,
@@ -233,25 +280,16 @@ class DeployerBasicMode {
       assetID: asa,
       payFlags: {}
     });
+    let asaId = 0;
     try {
-      const asaId = this.getASAInfo(asa).assetIndex;
-      await this.algoOp.optInAccountToASA(
-        asa,
-        asaId,
-        this._getAccount(accountName),
-        flags);
+      asaId = this.getASAInfo(asa).assetIndex;
     } catch (error) {
-      console.log("Asset no found in checkpoints. Proceeding to check as Asset ID.");
       if (!Number(asa)) {
         throw Error("Please provide a valid Number to be used as ASA ID");
       }
-      const asaId = Number(asa);
-      await this.algoOp.optInAccountToASA(
-        asa,
-        asaId,
-        this._getAccount(accountName),
-        flags);
+      asaId = Number(asa);
     }
+    return this.algoOp.optInAccountToASA(asa, asaId, this._getAccount(accountName), flags);
   }
 
   /**
@@ -261,7 +299,11 @@ class DeployerBasicMode {
    * @param lsig logic signature
    * @param flags Transaction flags
    */
-  async optInLsigToASA (asa: string, lsig: LogicSigAccount, flags: wtypes.TxParams): Promise<void> {
+  optInLsigToASA (
+    asa: string,
+    lsig: LogicSigAccount,
+    flags: wtypes.TxParams
+  ): Promise<void> {
     this.assertCPNotDeleted({
       type: wtypes.TransactionType.OptInASA,
       sign: wtypes.SignType.LogicSignature,
@@ -270,17 +312,16 @@ class DeployerBasicMode {
       assetID: asa,
       payFlags: {}
     });
+    let asaId = 0;
     try {
-      const asaId = this.getASAInfo(asa).assetIndex;
-      await this.algoOp.optInLsigToASA(asa, asaId, lsig, flags);
+      asaId = this.getASAInfo(asa).assetIndex;
     } catch (error) {
-      console.log("Asset no found in checkpoints. Proceeding to check as Asset ID.");
       if (!Number(asa)) {
         throw Error("Please provide a valid Number to be used as ASA ID");
       }
-      const asaId = Number(asa);
-      await this.algoOp.optInLsigToASA(asa, asaId, lsig, flags);
+      asaId = Number(asa);
     }
+    return this.algoOp.optInLsigToASA(asa, asaId, lsig, flags);
   }
 
   /**
@@ -291,11 +332,12 @@ class DeployerBasicMode {
    * @param payFlags Transaction flags
    * @param flags Optional parameters to SSC (accounts, args..)
    */
-  async optInAccountToApp (
+  optInAccountToApp (
     sender: rtypes.Account,
     appID: number,
     payFlags: wtypes.TxParams,
-    flags: rtypes.AppOptionalFlags): Promise<void> {
+    flags: rtypes.AppOptionalFlags
+  ): Promise<void> {
     this.assertCPNotDeleted({
       type: wtypes.TransactionType.OptInToApp,
       sign: wtypes.SignType.SecretKey,
@@ -303,7 +345,7 @@ class DeployerBasicMode {
       appID: appID,
       payFlags: {}
     });
-    await this.algoOp.optInAccountToApp(sender, appID, payFlags, flags);
+    return this.algoOp.optInAccountToApp(sender, appID, payFlags, flags);
   }
 
   /**
@@ -314,11 +356,12 @@ class DeployerBasicMode {
    * @param payFlags Transaction flags
    * @param flags Optional parameters to SSC (accounts, args..)
    */
-  async optInLsigToApp (
+  optInLsigToApp (
     appID: number,
     lsig: LogicSigAccount,
     payFlags: wtypes.TxParams,
-    flags: rtypes.AppOptionalFlags): Promise<void> {
+    flags: rtypes.AppOptionalFlags
+  ): Promise<void> {
     this.assertCPNotDeleted({
       type: wtypes.TransactionType.OptInToApp,
       sign: wtypes.SignType.LogicSignature,
@@ -327,18 +370,7 @@ class DeployerBasicMode {
       appID: appID,
       payFlags: {}
     });
-    await this.algoOp.optInLsigToApp(appID, lsig, payFlags, flags);
-  }
-
-  /**
-   * Returns ASCCache (with compiled code)
-   * @param name: Smart Contract filename (must be present in assets folder)
-   * @param force: if force is true file will be compiled for sure, even if it's checkpoint exist
-   * @param scTmplParams: scTmplParams: Smart contract template parameters
-   *     (used only when compiling PyTEAL to TEAL)
-   */
-  async ensureCompiled (name: string, force?: boolean, scTmplParams?: SCParams): Promise<ASCCache> {
-    return await this.algoOp.ensureCompiled(name, force, scTmplParams);
+    return this.algoOp.optInLsigToApp(appID, lsig, payFlags, flags);
   }
 
   /**
@@ -359,10 +391,9 @@ class DeployerBasicMode {
       res = key ? this.asa.get(key) : undefined;
     }
     if (res?.deleted === true) {
-      throw new BuilderError(
-        ERRORS.GENERAL.ASSET_DELETED, {
-          asset: asset
-        });
+      throw new BuilderError(ERRORS.GENERAL.ASSET_DELETED, {
+        asset: asset
+      });
     }
   }
 
@@ -378,11 +409,10 @@ class DeployerBasicMode {
   private assertAppExist (appID: number): void {
     const key = this.checkpoint.getAppCheckpointKeyFromIndex(appID);
     const res = key ? this.checkpoint.getAppfromCPKey(key) : undefined;
-    if (res?.deleted === true) {
-      throw new BuilderError(
-        ERRORS.GENERAL.APP_DELETED, {
-          app: appID
-        });
+    if (res?.deleted) {
+      throw new BuilderError(ERRORS.GENERAL.APP_DELETED, {
+        app: appID
+      });
     }
   }
 
@@ -451,10 +481,9 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
       return;
     }
     if (found) {
-      throw new BuilderError(
-        ERRORS.BUILTIN_TASKS.DEPLOYER_METADATA_ALREADY_PRESENT, {
-          metadataKey: key
-        });
+      throw new BuilderError(ERRORS.BUILTIN_TASKS.DEPLOYER_METADATA_ALREADY_PRESENT, {
+        metadataKey: key
+      });
     }
     this.cpData.putMetadata(this.networkName, key, value);
   }
@@ -466,10 +495,9 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
   assertNoAsset (name: string): void {
     if (this.isDefined(name)) {
       this.persistCP();
-      throw new BuilderError(
-        ERRORS.BUILTIN_TASKS.DEPLOYER_ASSET_ALREADY_PRESENT, {
-          assetName: name
-        });
+      throw new BuilderError(ERRORS.BUILTIN_TASKS.DEPLOYER_ASSET_ALREADY_PRESENT, {
+        assetName: name
+      });
     }
   }
 
@@ -516,10 +544,9 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
     if (asaDef === undefined) {
       this.persistCP();
 
-      throw new BuilderError(
-        ERRORS.BUILTIN_TASKS.DEPLOYER_ASA_DEF_NOT_FOUND, {
-          asaName: name
-        });
+      throw new BuilderError(ERRORS.BUILTIN_TASKS.DEPLOYER_ASA_DEF_NOT_FOUND, {
+        asaName: name
+      });
     }
     return await this.deployASADef(name, asaDef, flags);
   }
@@ -541,7 +568,12 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
     let asaInfo = {} as rtypes.ASAInfo;
     try {
       asaInfo = await this.algoOp.deployASA(
-        name, asaDef, flags, this.accountsByName, this.txWriter);
+        name,
+        asaDef,
+        flags,
+        this.accountsByName,
+        this.txWriter
+      );
     } catch (error) {
       this.persistCP();
 
@@ -557,7 +589,8 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
         asaDef,
         flags,
         this.accountsByName,
-        asaInfo.assetIndex);
+        asaInfo.assetIndex
+      );
     } catch (error) {
       this.persistCP();
 
@@ -575,8 +608,12 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
    * @param payFlags - as per SPEC
    * @param scTmplParams: Smart contract template parameters (used only when compiling PyTEAL to TEAL)
    */
-  async fundLsig (name: string, flags: FundASCFlags,
-    payFlags: wtypes.TxParams, scTmplParams?: SCParams): Promise<void> {
+  async fundLsig (
+    name: string,
+    flags: FundASCFlags,
+    payFlags: wtypes.TxParams,
+    scTmplParams?: SCParams
+  ): Promise<void> {
     try {
       await this.algoOp.fundLsig(name, flags, payFlags, this.txWriter, scTmplParams);
     } catch (error) {
@@ -595,7 +632,9 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
    * contract(optional in case of contract account)
    */
   async _mkLsig (
-    name: string, scTmplParams?: SCParams, signer?: rtypes.Account
+    name: string,
+    scTmplParams?: SCParams,
+    signer?: rtypes.Account
   ): Promise<LsigInfo> {
     this.assertNoAsset(name);
     let lsigInfo = {} as any;
@@ -635,8 +674,10 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
    *     (used only when compiling PyTEAL to TEAL)
    */
   async mkDelegatedLsig (
-    name: string, signer: rtypes.Account,
-    scTmplParams?: SCParams): Promise<LsigInfo> {
+    name: string,
+    signer: rtypes.Account,
+    scTmplParams?: SCParams
+  ): Promise<LsigInfo> {
     return await this._mkLsig(name, scTmplParams, signer);
   }
 
@@ -658,7 +699,7 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
    * @param scTmplParams: scTmplParams: Smart contract template parameters
    *     (used only when compiling PyTEAL to TEAL)
    * @param appName name of the app to deploy. This name (if passed) will be used as
-   * the checkpoint "key", and app information will be stored agaisnt this name
+   * the checkpoint "key", and app information will be associated with this name
    */
   async deployApp (
     approvalProgram: string,
@@ -666,14 +707,21 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
     flags: rtypes.AppDeploymentFlags,
     payFlags: wtypes.TxParams,
     scTmplParams?: SCParams,
-    appName?: string): Promise<rtypes.SSCInfo> {
-    const name = appName ?? (approvalProgram + "-" + clearProgram);
+    appName?: string
+  ): Promise<rtypes.SSCInfo> {
+    const name = appName ?? approvalProgram + "-" + clearProgram;
 
     this.assertNoAsset(name);
     let sscInfo = {} as rtypes.SSCInfo;
     try {
       sscInfo = await this.algoOp.deployApp(
-        approvalProgram, clearProgram, flags, payFlags, this.txWriter, scTmplParams);
+        approvalProgram,
+        clearProgram,
+        flags,
+        payFlags,
+        this.txWriter,
+        scTmplParams
+      );
     } catch (error) {
       this.persistCP();
 
@@ -696,7 +744,7 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
    * @param scTmplParams: scTmplParams: Smart contract template parameters
    *     (used only when compiling PyTEAL to TEAL)
    * @param appName name of the app to deploy. This name (if passed) will be used as
-   * the checkpoint "key", and app information will be stored agaisnt this name
+   * the checkpoint "key", and app information will be associated with this name
    */
   async updateApp (
     sender: algosdk.Account,
@@ -717,12 +765,20 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
       appID: appID,
       payFlags: {}
     });
-    const cpKey = appName ?? (newApprovalProgram + "-" + newClearProgram);
+    const cpKey = appName ?? newApprovalProgram + "-" + newClearProgram;
 
     let sscInfo = {} as rtypes.SSCInfo;
     try {
-      sscInfo = await this.algoOp.updateApp(sender, payFlags, appID,
-        newApprovalProgram, newClearProgram, flags, this.txWriter, scTmplParams);
+      sscInfo = await this.algoOp.updateApp(
+        sender,
+        payFlags,
+        appID,
+        newApprovalProgram,
+        newClearProgram,
+        flags,
+        this.txWriter,
+        scTmplParams
+      );
     } catch (error) {
       this.persistCP();
 
@@ -751,10 +807,9 @@ export class DeployerRunMode extends DeployerBasicMode implements Deployer {
 
   assertNoAsset (name: string): void {
     if (this.isDefined(name)) {
-      throw new BuilderError(
-        ERRORS.BUILTIN_TASKS.DEPLOYER_ASSET_ALREADY_PRESENT, {
-          assetName: name
-        });
+      throw new BuilderError(ERRORS.BUILTIN_TASKS.DEPLOYER_ASSET_ALREADY_PRESENT, {
+        assetName: name
+      });
     }
   }
 
@@ -798,15 +853,22 @@ export class DeployerRunMode extends DeployerBasicMode implements Deployer {
     });
   }
 
-  async fundLsig (_name: string, _flags: FundASCFlags,
-    _payFlags: wtypes.TxParams, _scInitParams?: unknown): Promise<LsigInfo> {
+  async fundLsig (
+    _name: string,
+    _flags: FundASCFlags,
+    _payFlags: wtypes.TxParams,
+    _scInitParams?: unknown
+  ): Promise<LsigInfo> {
     throw new BuilderError(ERRORS.BUILTIN_TASKS.DEPLOYER_EDIT_OUTSIDE_DEPLOY, {
       methodName: "fundLsig"
     });
   }
 
-  async mkDelegatedLsig (_name: string, _signer: rtypes.Account,
-    _scInitParams?: unknown): Promise<LsigInfo> {
+  async mkDelegatedLsig (
+    _name: string,
+    _signer: rtypes.Account,
+    _scInitParams?: unknown
+  ): Promise<LsigInfo> {
     throw new BuilderError(ERRORS.BUILTIN_TASKS.DEPLOYER_EDIT_OUTSIDE_DEPLOY, {
       methodName: "delegatedLsig"
     });
@@ -818,7 +880,8 @@ export class DeployerRunMode extends DeployerBasicMode implements Deployer {
     flags: rtypes.AppDeploymentFlags,
     payFlags: wtypes.TxParams,
     scInitParam?: unknown,
-    appName?: string): Promise<rtypes.SSCInfo> {
+    appName?: string
+  ): Promise<rtypes.SSCInfo> {
     throw new BuilderError(ERRORS.BUILTIN_TASKS.DEPLOYER_EDIT_OUTSIDE_DEPLOY, {
       methodName: "deployApp"
     });
@@ -855,9 +918,14 @@ export class DeployerRunMode extends DeployerBasicMode implements Deployer {
       payFlags: {}
     });
     return await this.algoOp.updateApp(
-      sender, payFlags, appID,
-      newApprovalProgram, newClearProgram,
-      flags, this.txWriter, scTmplParams
+      sender,
+      payFlags,
+      appID,
+      newApprovalProgram,
+      newClearProgram,
+      flags,
+      this.txWriter,
+      scTmplParams
     );
   }
 }