+++ packages/runtime/src/interpreter/opcode-list.ts
@@ -12,6 +12,7 @@ import { RuntimeError } from "../errors/runtime-errors";
 import { compareArray } from "../lib/compare";
 import {
   ALGORAND_MAX_LOGS_COUNT, ALGORAND_MAX_LOGS_LENGTH,
+  AppParamDefined,
   AssetParamMap, GlobalFields, MathOp,
   MAX_CONCAT_SIZE, MAX_INNER_TRANSACTIONS,
   MAX_INPUT_BYTE_LEN, MAX_OUTPUT_BYTE_LEN,
@@ -20,7 +21,7 @@ import {
 } from "../lib/constants";
 import { parseEncodedTxnToExecParams, setInnerTxField } from "../lib/itxn";
 import {
-  assertLen, assertOnlyDigits, bigEndianBytesToBigInt, bigintToBigEndianBytes, convertToBuffer,
+  assertLen, assertNumber, assertOnlyDigits, bigEndianBytesToBigInt, bigintToBigEndianBytes, convertToBuffer,
   convertToString, getEncoding, parseBinaryStrToBigInt
 } from "../lib/parsing";
 import { Stack } from "../lib/stack";
@@ -1610,8 +1611,8 @@ export class Global extends Op {
         break;
       }
       case 'CreatorAddress': {
-        const appID = this.interpreter.runtime.ctx.tx.apid;
-        const app = this.interpreter.getApp(appID as number, this.line);
+        const appID = this.interpreter.runtime.ctx.tx.apid as number;
+        const app = this.interpreter.getApp(appID, this.line);
         result = decodeAddress(app.creator).publicKey;
         break;
       }
@@ -2147,8 +2148,8 @@ export class Int extends Op {
     if (intConst !== undefined) {
       uint64 = BigInt(intConst);
     } else {
-      assertOnlyDigits(args[0], line);
-      uint64 = BigInt(args[0]);
+      const val = assertNumber(args[0], line);
+      uint64 = BigInt(val);
     }
 
     this.checkOverflow(uint64, line, MAX_UINT64);
@@ -3671,11 +3672,13 @@ export class Uncover extends Op {
     this.assertMinStackLen(stack, this.nthInStack + 1, this.line);
 
     const temp = [];
-    for (let count = 1; count < this.nthInStack; ++count) {
+    for (let count = 0; count < this.nthInStack; ++count) {
       temp.push(stack.pop());
     }
+
     const deepValue = stack.pop();
-    for (let i = this.nthInStack - 2; i >= 0; --i) {
+
+    for (let i = this.nthInStack - 1; i >= 0; --i) {
       stack.push(temp[i]);
     }
     stack.push(deepValue);
@@ -4232,3 +4235,120 @@ export class Log extends Op {
     txReceipt.logs.push(convertToString(logByte));
   }
 }
+
+// bitlen interprets arrays as big-endian integers, unlike setbit/getbit
+// stack = [..., any]
+// push to stack = [..., bitlen]
+export class BitLen extends Op {
+  readonly line: number;
+
+  /**
+   * Asserts 0 arguments are passed.
+   * @param args Expected arguments: [] // none
+   * @param line line number in TEAL file
+   * @param interpreter interpreter object
+  */
+  constructor (args: string[], line: number) {
+    super();
+    this.line = line;
+    assertLen(args.length, 0, line);
+  };
+
+  execute (stack: Stack<StackElem>): void {
+    this.assertMinStackLen(stack, 1, this.line);
+    const value = stack.pop();
+
+    let bitlen = 0;
+
+    if (typeof value === "bigint") {
+      bitlen = (value === 0n) ? 0 : value.toString(2).length;
+    } else {
+      // value is Uint8 => one element have 8 bits.
+      // => bitlen = 8 * value.length - 1 + bitlen(first element)
+      if (value.length > 0) {
+        bitlen = (value.length - 1) * 8;
+        bitlen += value[0].toString(2).length;
+      }
+    }
+
+    stack.push(BigInt(bitlen));
+  }
+}
+
+// get App Params Information
+// push to stack [...stack, value(bigint/bytes), did_exist]
+// NOTE: if app doesn't exist, then did_exist = 0, value = 0
+export class AppParamsGet extends Op {
+  readonly interpreter: Interpreter;
+  readonly line: number;
+  readonly field: string;
+  /**
+   * Asserts 1 arguments are passed.
+   * @param args Expected arguments: [] // none
+   * @param line line number in TEAL file
+   * @param interpreter interpreter object
+   */
+  constructor (args: string[], line: number, interpreter: Interpreter) {
+    super();
+    this.line = line;
+    this.interpreter = interpreter;
+    assertLen(args.length, 1, line);
+
+    if (!AppParamDefined[interpreter.tealVersion].has(args[0])) {
+      throw new RuntimeError(RUNTIME_ERRORS.TEAL.UNKNOWN_APP_FIELD, {
+        field: args[0],
+        line: line,
+        tealV: interpreter.tealVersion
+      });
+    }
+
+    this.field = args[0];
+  }
+
+  execute (stack: Stack<StackElem>): void {
+    this.assertMinStackLen(stack, 1, this.line);
+
+    const appID = this.assertBigInt(stack.pop(), this.line);
+
+    if (this.interpreter.runtime.ctx.state.globalApps.has(Number(appID))) {
+      let value: StackElem = 0n;
+      const appDef = this.interpreter.getApp(Number(appID), this.line);
+      switch (this.field) {
+        case "AppApprovalProgram":
+          value = parsing.stringToBytes(appDef["approval-program"]);
+          break;
+        case "AppClearStateProgram":
+          value = parsing.stringToBytes(appDef["clear-state-program"]);
+          break;
+        case "AppGlobalNumUint":
+          value = BigInt(appDef["global-state-schema"].numUint);
+          break;
+        case "AppGlobalNumByteSlice":
+          value = BigInt(appDef["global-state-schema"].numByteSlice);
+          break;
+        case "AppLocalNumUint":
+          value = BigInt(appDef["local-state-schema"].numUint);
+          break;
+        case "AppLocalNumByteSlice":
+          value = BigInt(appDef["local-state-schema"].numByteSlice);
+          break;
+        case "AppExtraProgramPages":
+          // only return default number extra program pages in runtime
+          // should fix it in future.
+          value = 1n;
+          break;
+        case "AppCreator":
+          value = decodeAddress(appDef.creator).publicKey;
+          break;
+        case "AppAddress":
+          value = decodeAddress(getApplicationAddress(appID)).publicKey;
+      };
+
+      stack.push(value);
+      stack.push(1n);
+    } else {
+      stack.push(0n);
+      stack.push(0n);
+    }
+  };
+};