+++ packages/web/test/src/types.ts
@@ -7,7 +7,7 @@ describe("Transaction And Sign interface", () => {
   it("should return false if transaction object is not SDK transaction", () => {
     const param: unknown = {
       transaction: { name: "AA" },
-      sign: { sign: "sd" }
+      sign: { sign: "sd" },
     };
 
     assert.isFalse(isSDKTransactionAndSign(param));
@@ -15,7 +15,7 @@ describe("Transaction And Sign interface", () => {
 
   it("should return false if sign is not in transaction", () => {
     const param: unknown = {
-      transaction: txObject
+      transaction: txObject,
     };
 
     assert.isFalse(isSDKTransactionAndSign(param));
@@ -24,7 +24,7 @@ describe("Transaction And Sign interface", () => {
   it("should return true if sign and transaction is present", () => {
     const param: unknown = {
       transaction: txObject,
-      sign: { sign: SignType.SecretKey, fromAccount: [] }
+      sign: { sign: SignType.SecretKey, fromAccount: [] },
     };
 
     assert.isTrue(isSDKTransactionAndSign(param));